requirements.txt
uvicorn[standard]>=0.34.0,<0.34.2
fastapi[all]>=0.115.0,<0.115.12
python-dotenv==1.1.0
pydantic==2.11.4
pyyaml==6.0.2